move more implementation details to array_domain.cpp
